new task

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -5,7 +5,7 @@
 int main(void)
 {
 	int i;
-	long long int fibb[50];
+	long int fibb[50];
 	long int sum;
 
 	sum = 0;
